Feat: Organizar o projeto backend (#13)
* feat: organiza as pasta do projeto backend

* feat: adiciona o changeset
