+++ backend/src/config/database.js
@@ -1,0 +1,13 @@
+const { Pool } = require("pg");
+
+require("dotenv").config();
+
+const pool = new Pool({
+  user: process.env.DATABASE_USER,
+  host: process.env.DATABASE_HOST,
+  database: process.env.DATABASE_SCHEMA,
+  password: process.env.DATABASE_PASSWORD,
+  port: process.env.DATABASE_PORT,
+});
+
+module.exports = pool;
