+++ backend/src/constants/messages.js
@@ -1,0 +1,10 @@
+module.exports.messages = {
+  success: {
+    userRegistered: "Usuário cadastrado no QA Solar com sucesso!",
+    loginSuccess: "Usuário logado com sucesso!",
+  },
+  error: {
+    userRegistrationFailed: "Erro ao cadastrar o usuário.",
+    loginFailure: "Não foi possivel fazer login com este usuário!",
+  },
+};
