--- backend/src/controllers.js
@@ -1,25 +1,0 @@
-const { PrismaClient } = require("@prisma/client");
-const prisma = new PrismaClient();
-
-const createUser = async (req, res) => {
-  try {
-    const { name, email, password } = req.body;
-    const user = await prisma.user.create({
-      data: { name, email, password },
-    });
-    res.status(201).json(user);
-  } catch (error) {
-    res.status(400).json({ error: "Erro ao criar usuário" });
-  }
-};
-
-const getUsers = async (req, res) => {
-  try {
-    const users = await prisma.user.findMany();
-    res.status(200).json(users);
-  } catch (error) {
-    res.status(500).json({ error: "Erro ao buscar usuários" });
-  }
-};
-
-module.exports = { createUser, getUsers };
