+++ backend/src/server.js
@@ -1,0 +1,7 @@
+const app = require("./app");
+
+const port = process.env.PORT || 3001;
+
+app.listen(port, () => {
+  console.log(`Servidor rodando em http://localhost:${port}`);
+});
